adjust folder position
